Fixed warnings in MSVC. Removed unused sim-math includes until cross platform solution is found. Implemented SDL InterProcessMutex as a shared writable file. Fixed inconsistent newlines.

+++ ZeroLibraries/AudioEngine/Filters.cpp
@@ -1964,7 +1964,7 @@ namespace Audio
       break;
     }
 
-    mCurrentTime += cSystemTimeIncrement;
+    mCurrentTime += (float)cSystemTimeIncrement;
     mLastAmplitude = amplitude;
 
     return amplitude;
